backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
SQLAlchemy==2.0.23
mysql-connector-python==8.2.0
pymysql==1.1.0
psycopg2-binary==2.9.9
cx-Oracle==8.3.0
pyodbc==5.0.1
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.2
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.5
marshmallow==3.20.1
pytest==7.4.3 